Первый раздел презентера осуществлён
